+++ src/MyFiles.java
@@ -39,7 +39,7 @@ class MyFiles {
 			// populate grid
 			for (int i=0; i<rows; i++) {
 				for (int j=0; j<cols; j++) {
-					grid[i][j] = new PointElevation(inScanner.nextDouble());
+					grid[i][j] = new PointElevation(inScanner.nextFloat());
 				}
 			}
